reduce boilerplate, use utils
