+++ shared/ui/Stream/NRQL/NRQLResultsLine.tsx
@@ -11,29 +11,20 @@ import {
 } from "recharts";
 import { NRQLResult } from "@codestream/protocols/agent";
 import { isEmpty as _isEmpty } from "lodash-es";
-import { ColorsHash, Colors } from "./utils";
+import {
+	ColorsHash,
+	Colors,
+	truncate,
+	fillNullValues,
+	getUniqueDataKeyAndFacetValues,
+	formatXAxisTime,
+} from "./utils";
 import { EventTypeTooltip } from "./EventTypeTooltip";
 import { EventTypeLegend } from "./EventTypeLegend";
 import { FacetLineTooltip } from "./FacetLineTooltip";
 import Tooltip from "../Tooltip";
 
 export const LEFT_MARGIN_ADJUST_VALUE = 25;
-
-const formatXAxisTime = time => {
-	const date = new Date(time * 1000);
-	return `${date.toLocaleTimeString()}`;
-};
-
-const getUniqueDataKeyAndFacetValues = (results, facet) => {
-	const result = results ? results[0] : undefined;
-
-	const defaultFilterKeys = ["beginTimeSeconds", "endTimeSeconds", "facet"];
-	const filterKeys = defaultFilterKeys.concat(facet);
-
-	const dataKeys = Object.keys(result || {}).filter(key => !filterKeys.includes(key));
-	const uniqueFacetValues = [...new Set(results.map(obj => obj.facet))];
-	return { dataKeys, uniqueFacetValues };
-};
 
 const formatResultsForLineChart = (originalArray, uniqueFacets, dataKeys) => {
 	const groupedByEndTime = {};
@@ -59,27 +50,6 @@ const formatResultsForLineChart = (originalArray, uniqueFacets, dataKeys) => {
 	}));
 
 	return fillNullValues(newArray);
-};
-
-const fillNullValues = array => {
-	array.forEach((obj, i) => {
-		Object.keys(obj).forEach(key => {
-			if (key !== "endTimeSeconds" && obj[key] === null) {
-				let j = i - 1;
-				while (j >= 0 && array[j][key] === null) j--;
-				obj[key] = j >= 0 ? array[j][key] : 0;
-			}
-		});
-	});
-	return array.filter(obj =>
-		Object.keys(obj).some(key => key !== "endTimeSeconds" && obj[key] !== undefined)
-	);
-};
-
-const truncate = (str: string, max: number) => {
-	if (!str) return str;
-	if (str.length >= max) return `${str.substring(0, max - 1)}${"\u2026"}`;
-	return str;
 };
 
 interface NRQLResultsLineProps {
